Add features and rebase kv store on simple resolver

+++ packages/graphweaver-simple-resolver/src/provider.ts
@@ -8,33 +8,39 @@ import { Options } from "./types";
 
 export class SimpleProvider<D, G, Ctx> implements BackendProvider<D, G> {
   public readonly backendId: string;
+  protected create: Options["create"];
   protected read: Options["read"];
-  protected write: Options["write"];
+  protected update: Options["update"];
+  protected remove: Options["remove"];
+
   public context: Ctx | undefined = undefined;
   private init: Promise<Ctx>;
 
-  constructor({ name, read, write, init }: Options<D>) {
+  constructor({ name, create, read, update, remove, init }: Options<D>) {
     this.backendId = `SimpleResolver`;
+    this.create = create;
     this.read = read;
-    this.write = write;
+    this.update = update;
+    this.remove = remove;
     this.init = new Promise(async (resolve) => {
-      console.log("Init...");
       this.context = await init?.();
-      console.log("Context set", this.context);
       resolve();
     });
   }
 
   async find(filter: Filter<G>, pagination?: PaginationOptions): Promise<D[]> {
     await this.init;
-    return this.read(this.context, filter, pagination) ?? [];
+    const result = await this.read(this.context, filter, pagination);
+    return Array.isArray(result) ? result : [result];
   }
 
   async findOne(filter: Filter<G>): Promise<D | null> {
     await this.init;
+    console.log("findOne");
     const result = await this.read(this.context, filter);
-    console.log("findOne result", result);
-    return result?.[0] || null;
+    console.log("Found", result);
+    const oneResult = Array.isArray(result) ? result?.[0] : result;
+    return oneResult || null;
   }
 
   findByRelatedId(
@@ -47,28 +53,38 @@ export class SimpleProvider<D, G, Ctx> implements BackendProvider<D, G> {
   }
 
   async updateOne(id: string, updateArgs: Partial<G>): Promise<D> {
+    if (!this.update) throw new Error("update not available");
     await this.init;
-    return this.write(this.context, updateArgs, id);
+    return this.update(this.context, id, updateArgs);
   }
 
-  updateMany(entities: Partial<G>[]): Promise<D[]> {
+  async updateMany(entities: Partial<G>[]): Promise<D[]> {
+    if (!this.update) throw new Error("update not available");
+    await this.init;
     throw new Error("Not implemented: updateMany");
   }
 
   async createOne(entity: Partial<G>): Promise<D> {
+    if (!this.create) throw new Error("create not available");
     await this.init;
-    return this.write(this.context, entity);
+    return this.create(this.context, entity);
   }
 
   async createMany(entities: Partial<G>[]): Promise<D[]> {
+    if (!this.create) throw new Error("create not available");
+    await this.init;
     throw new Error("Not implemented: createMany");
   }
 
-  createOrUpdateMany(entities: Partial<G>[]): Promise<D[]> {
+  async createOrUpdateMany(entities: Partial<G>[]): Promise<D[]> {
+    if (!this.update || !this.create)
+      throw new Error("create/update not available");
     throw new Error("Not implemented: createOrUpdateMany");
   }
 
-  deleteOne(filter: Filter<G>): Promise<boolean> {
+  async deleteOne(filter: Filter<G>): Promise<boolean> {
+    if (!this.remove) throw new Error("delete not available");
+    await this.init;
     throw new Error("Not implemented: deleteOne");
   }
